+++ packages/storage/src/index.ts
@@ -1,0 +1,9 @@
+import { createStorage } from './createStorage';
+
+let globalStorage: Storage;
+
+export const getGlobalStorage = () => {
+  globalStorage ||= createStorage();
+
+  return globalStorage;
+};