+++ packages/wallet-engine/src/engine/wallet.ts
@@ -1,33 +1,57 @@
 import { createScaffoldMiddleware, createAsyncMiddleware } from 'json-rpc-engine';
 
 import { Engine } from './engine';
-import { ChainConfig, Account } from '../types';
+import { ChainConfig } from '../types';
 
 export type WalletEngineOptions = {
-  getAccounts: () => Account[];
+  getPrivateKey: () => string | undefined;
   chainConfig: ChainConfig;
 };
 
-export const createWalletEngine = ({ getAccounts = () => [], chainConfig }: WalletEngineOptions) => {
+export const createWalletEngine = ({ chainConfig, getPrivateKey }: WalletEngineOptions) => {
   const engine = new Engine();
 
   engine.push(
     createScaffoldMiddleware({
-      wallet_getProviderState: createAsyncMiddleware(async (req, res) => {
-        res.result = {
-          accounts: getAccounts().map((account) => account.address),
-          chainId: chainConfig.chainId,
-          isUnlocked: true,
-          networkVersion: chainConfig.chainId,
-        };
-      }),
-
       wallet_sendDomainMetadata: createAsyncMiddleware(async (req, res) => {
         res.result = true;
       }),
 
-      wallet_accounts: createAsyncMiddleware(async (req, res) => {
-        res.result = getAccounts();
+      eth_chainId: createAsyncMiddleware(async (req, res) => {
+        res.result = chainConfig.chainId;
+      }),
+
+      net_version: createAsyncMiddleware(async (req, res) => {
+        res.result = chainConfig.chainId;
+      }),
+
+      wallet_accounts: createAsyncMiddleware(async (req, res, next) => {
+        res.result = [];
+
+        return getPrivateKey() ? next() : undefined;
+      }),
+
+      wallet_updateAccounts: createAsyncMiddleware(async (req, res, next) => {
+        res.result = [];
+
+        return getPrivateKey() ? next() : undefined;
+      }),
+
+      wallet_getProviderState: createAsyncMiddleware(async (req, res, next) => {
+        res.result = {
+          accounts: [],
+          chainId: chainConfig.chainId,
+          isUnlocked: true,
+          networkVersion: chainConfig.chainId,
+        };
+
+        return getPrivateKey() ? next() : undefined;
+      }),
+
+      eth_requestAccounts: createAsyncMiddleware(async (req, res, next) => {
+        res.result = [];
+
+        return getPrivateKey() ? next() : undefined;
       }),
     }),
   );