+++ src/routes/Authorize/index.ts
@@ -1,3 +1,4 @@
+export { default as Authorize } from './Authorize';
 export { default as AuthorizeClose } from './AuthorizeClose';
 export { default as AuthorizeRedirect } from './AuthorizeRedirect';
 export { IntroRoute } from './IntroRoute';