+++ CHANGELOG.md
@@ -3,7 +3,18 @@
 ### vNext
 
 - Social login with redirect
+- Use safe `localStorage` wrapper
+- Fix `Sign Out` from the wallet widget error
+- Send only ETH account address to applications tracking endpoints
+
+### v1.22.0
+
 - Disable Biconomy for CERE ERC20 transfer
+- Application bundle optimization to improve load speed
+- Add CERE ERC20 token to default assets
+- Disable OPT code auto-capitalization
+- Optimize login flow to decrease time to login
+- Increase assets balance pooling interval to 10 seconds
 
 ### v1.21.0